requirements.txt
Flask==2.2.2
google-cloud-datastore==2.10.0
protobuf==3.20.3
python-jose
flask-cors
six
python-dotenv
requests
authlib
